+++ frontend/src/pages/Login/index.ts
@@ -2,9 +2,8 @@ import { API_URL } from '@/config/config';
 import i18next from '@/config/i18n';
 import { Page } from '@/core/Page';
 import CommonLayout from '@/layouts/common/index';
-import { updateActiveLanguageButton } from '@/models/Lang/repository';
 import { updateLanguage, updateOnlineStatus } from '@/models/User/repository';
-import { initLanguageSwitchers } from '@/utils/language';
+import { initLanguageSwitchers, updateActiveLanguageButton } from '@/utils/language';
 
 const updatePageContent = (): void => {
   const loginTitle = document.querySelector('.login-container h1');
